Language.Haskell.TypeTree

GHCi setup

>>> :set -XTemplateHaskell -XTypeFamilies -XGADTs

Usage

Basic usage

ttReify allows you to build a Tree containing type information for each field of any given datatype, which can then be examined if you want to, for example, generate class instances for a deeply nested datatype. (The idea for this package came about when I was trying to figure out the easiest way to generate several dozen instances for Cabal's GenericPackageDescription.)

Plain constructors

>>> data Foo a = Foo { field1 :: Either a Int }
>>> putStr $(ttDescribe ''Foo)
Ghci4.Foo a_0
|
`- Data.Either.Either a_0 GHC.Types.Int
   |
   +- $a_0
   |
   `- GHC.Types.Int

GADTs

>>> data MyGADT a where Con1 :: String -> MyGADT String; Con2 :: Int -> MyGADT [Int]
>>> putStr $(ttDescribe ''MyGADT)
Ghci8.MyGADT
|
+- GHC.Base.String
|  |
|  `- GHC.Types.[] GHC.Types.Char
|     |
|     `- GHC.Types.Char
|
+- GHC.Base.String
|  |
|  `- GHC.Types.[] GHC.Types.Char
|     |
|     `- GHC.Types.Char
|
+- GHC.Types.Int
|
`- GHC.Types.[] GHC.Types.Int
   |
   `- GHC.Types.Int

Constructor return types are treated as another field in this case.

Data family instances

>>> class Foo a where data Bar a :: * -> *
>>> instance Foo Int where data Bar Int a = IntBar { bar :: Maybe (Int, a) }
>>> type IntBar = Bar Int
>>> putStr $(ttDescribe ''IntBar)
Ghci16.IntBar
|
`- Ghci12.Bar GHC.Types.Int a_0
   |
   `- GHC.Base.Maybe (GHC.Types.Int, a_0)
      |
      `- GHC.Tuple.(,) GHC.Types.Int a_0
         |
         +- GHC.Types.Int
         |
         `- $a_0

Recursive datatypes

>>> data Foo a = Foo { a :: Either Int (Bar a) }; data Bar b = Bar { b :: Either (Foo b) Int }
>>> putStr $(ttDescribe ''Foo)
Ghci20.Foo a_0
|
`- Data.Either.Either GHC.Types.Int (Ghci20.Bar a_0)
   |
   +- GHC.Types.Int
   |
   `- Ghci20.Bar a_0
      |
      `- Data.Either.Either (Ghci20.Foo a_0) GHC.Types.Int
         |
         +- <recursive Ghci20.Foo a_0>
         |
         `- GHC.Types.Int

Currently, if you want to instantiate type variables in the given datatype, the best way to do so is to define a type synonym:

>>> type MaybeChar = Maybe Char
>>> putStr $(ttDescribe ''MaybeChar)
Ghci24.MaybeChar
|
`- GHC.Base.Maybe GHC.Types.Char
   |
   `- GHC.Types.Char

...but future versions of this library may provide a more ergonomic way to pass type arguments.

Passing options

If needed, type-tree allows you to specify that primitive type constructors should be included in its output.

>>> data Baz = Baz { field :: [Int] }
>>> putStr $(ttDescribeOpts defaultOpts { expandPrim = True } ''Baz)
Ghci28.Baz
|
`- GHC.Types.[] GHC.Types.Int
   |
   `- GHC.Types.Int
      |
      `- GHC.Prim.Int#

Note that the function arrow (->), despite being a primitive type constructor, will always be included even with expandPrim = False, as otherwise you would never be able to get useful information out of a field with a function type.

You can also specify a set of names where type-tree should stop descending, if, for example, you have no desire to see String -> [] -> Char ad nauseam in your tree.

>>> data Bar = Bar (Either String [String])
>>> putStr $(ttDescribeOpts defaultOpts { terminals = S.fromList [''String] } ''Bar)
Ghci32.Bar
|
`- Data.Either.Either GHC.Base.String ([GHC.Base.String])
   |
   +- GHC.Base.String
   |
   `- GHC.Types.[] GHC.Base.String
      |
      `- GHC.Base.String

Producing trees

ttReify :: Name -> Q (Tree Leaf)

Build a "type tree" of the given datatype.

Concrete types will appear in the tree as ConL. Unbound variables will appear as VarL. If the datastructure is recursive, occurrences of the node after the first will be wrapped in Recursive.

ttReifyOpts :: ReifyOpts -> Name -> Q (Tree Leaf)

ttReify with the provided options.

ttLit :: Name -> ExpQ

Embed the produced tree as an expression.

ttLitOpts :: ReifyOpts -> Name -> ExpQ

ttLit with provided opts.

Debugging trees

ttDescribe :: Name -> ExpQ

Produces a string literal representing a type tree. Useful for debugging purposes.

ttDescribeOpts :: ReifyOpts -> Name -> ExpQ

ttDescribe with the given options.

Building graphs

type Key = (Name, [Type])

Some type and its arguments, as representable in a graph.

type Arity = Int

Type constructor arity.

ttEdges :: Name -> ExpQ

$(ttEdges ''Foo) :: [((Name, Arity), Key, [Key])]

$(ttEdges ''Foo) produces a list suitable for passing to graphFromEdges.

ttConnComp :: Name -> ExpQ

$(ttConnComp ''Foo) :: [SCC (Name, Arity)]

$(ttConnComp ''Foo) produces a topologically sorted list of the strongly connected components of the graph representing Foo.

Customizing trees

data Leaf

Constructors

ConL (Name, [Type])
ConL (name, xs) is a field with type ConT name, applied to types xs.

VarL (Name, [Type])
VarL (name, xs) is a field with type VarT name, applied to types xs.

Recursive Leaf
Recursive field.

data ReifyOpts

Constructors

ReifyOpts

Fields

expandPrim :: Bool

Descend into primitive type constructors?

terminals :: Set Name

If a name in this set is encountered, stop descending.

defaultOpts :: ReifyOpts

Default reify options.

defaultOpts = ReifyOpts
  { expandPrim = False
  , terminals = mempty
  }